Moved Footer related components to Footer folder + BrowseScanButtons is now generic
