+++ src/Detail.tsx
@@ -2,7 +2,7 @@ import { Skeleton } from "primereact/skeleton";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { Annotation } from "./components/Annotations/annotation";
-import { Footer } from "./components/Footer";
+import { Footer } from "./components/Footer/Footer";
 import { Mirador } from "./components/Mirador/Mirador";
 import { TextComponent } from "./components/Text/TextComponent";
 import { Broccoli } from "./model/Broccoli";
