+++ src/components/Annotations/AnnotationButtons.tsx
@@ -1,11 +1,12 @@
 import { projectConfigSelector, useProjectStore } from "../../stores/project";
+import { BrowseScanButtons } from "../Footer/BrowseScanButtons";
 
 export const AnnotationButtons = () => {
   const projectConfig = useProjectStore(projectConfigSelector);
   return (
     <>
       <projectConfig.components.AnnotationButtons />
-      <projectConfig.components.BrowseScanButtons />
+      {projectConfig.showPrevNextScanButtons ? <BrowseScanButtons /> : null}
     </>
   );
 };
